refactoring
changed all floats to double
changed function names (added set_ / get_) to the functionname
activationFunction fixes for gaussian
Performance Test example for developing
2D view of live training in XOR GUI
Changed the error handling system, no thrown errors anymore

+++ examples/Genetic_simple/inc/enviroment.h
@@ -50,8 +50,8 @@ class Enviroment{
         vector<bool>            _animalAliveList;
         vector<int>             _animalBackCountList;
         int                     _animalStartFood;
-        vector<float>           _animalScoreList;
-        float                   _averageScore;
+        vector<double>           _animalScoreList;
+        double                   _averageScore;
         unsigned int            _stillAlive;
         unsigned int            _cycles;
         unsigned int            _cyclesCounter;
